% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rwd_agg.R
\name{as_rwd_agg}
\alias{as_rwd_agg}
\alias{as_rwd_agg.data.frame}
\alias{as_rwd_agg.list}
\alias{as_rwd_agg.matrix}
\alias{as_rwd_agg.default}
\alias{as.rwd_agg}
\title{Coerce lists, matrices, and data.frames to RiverWare data aggregators}
\usage{
as_rwd_agg(x, ...)

\method{as_rwd_agg}{data.frame}(x, ...)

\method{as_rwd_agg}{list}(x, ...)

\method{as_rwd_agg}{matrix}(x, ...)

\method{as_rwd_agg}{default}(x, ...)
}
\arguments{
\item{x}{A list. Each element of the list must have the same length.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
S3 generic for coercing from lists, matrices, and data.frames to \link{rwd_agg}
objects.
}
