% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigrdf_to_rwtbl.R
\name{bigrdf_to_rwtbl}
\alias{bigrdf_to_rwtbl}
\alias{bigrdf_save}
\title{Convert a 'big' rdf file into a tibble-like object.}
\usage{
bigrdf_to_rwtbl(
  file,
  scenario = NA_character_,
  keep_cols = FALSE,
  add_ym = TRUE,
  status = TRUE,
  n_trace_per_chunk = 20
)

bigrdf_save(x, path, remove_uid = FALSE)
}
\arguments{
\item{file}{The relative or absolute rdf filename.}

\item{scenario}{An optional parameter, that if it is not \code{NULL} or \code{NA}
(default) will be added to the tibble as another variable. Coerced to a
character if it is not already a character.}

\item{keep_cols}{Either boolean, or a character vector of column names to
keep in the returned tibble. The values of \code{keep_cols} work as follows:
\itemize{
\item \code{FALSE} (default) only includes the defaults columns: \code{Timestep},
\code{TraceNumber}, \code{ObjectSlot}, and \code{Value}. \code{Scenario} is also returned if
\code{scenario} is specified.
\item \code{TRUE}, all columns are returned (see Details).
\item A character vector, e.g., \code{c("ObjectName", "Unit")}, allows the user to
include other columns that are not always required, in addition to the
"default" set of columns. If any of the values in \code{keep_cols} are not
found, a warning will post, but all other columns will be returned.
}}

\item{add_ym}{Boolean that controls whether or not \code{Year} and \code{Month} columns
are added to the returned tibble. If \code{TRUE} (default), they will be added,
and if \code{FALSE} they will not be added. They are constructed from the dates
in the \code{Timestep} column.}

\item{status}{If \code{TRUE}, print out information on status of processing the
rdf.}

\item{n_trace_per_chunk}{How many traces to process at a time. It is faster
to process more traces at a time, but you are more likely to run into
memory issues. It will be dependent on the overall structure of the rdf in
terms of number of variables and timesteps and will be up to the user to
determine an appropriate value in terms of speed vs. memory.}

\item{x}{Object returned by \code{bigrdf_to_rwtbl()}}

\item{path}{The file path where the user wants the bigrdf saved.}

\item{remove_uid}{In the temporary storage, a unique ID is included in the
FileSystemDataSet folder name. If this is \code{TRUE}, that UID is removed when
saving the data to \code{path}.}
}
\description{
\code{bigrdf_to_rwtbl()} converts an rdf into a tibble-like object. 'Big' rdf
files can cause memory issues when processing them, and keeping them stored
entirely in memory as tibbles. \code{bigrdf_to_rwtbl()} processes the rdf files
in chunks, specified as the number of traces per chunk (\code{n_trace_per_chunk}),
and saves the data as a temporary \link[arrow:Dataset]{arrow::FileSystemDataset}, where each
trace is saved as a single parquet file. This ensures there are no memory
issues when processing the rdf file.

Users can then save the \code{FileSystemDataSet} using \code{bigrdf_save()}, and/or
work with it from its temporary location in a tidyverse pipeline. Using
\code{\link[dplyr:compute]{dplyr::collect()}} will return a tibble in memory.
}
