% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{flowWeightedAvgAnnConc}
\alias{flowWeightedAvgAnnConc}
\title{Calculate the flow-weighted average annual concentration.}
\usage{
flowWeightedAvgAnnConc(mass, flow)
}
\arguments{
\item{mass}{A vector of one trace worth of data, or one year of one trace. Units should be in tons.}

\item{flow}{A vector of one trace worth of data, or one year of one trace. Units should be in acre-ft/month.}
}
\value{
A victor of yearly data of the flow-weighted average annual concentration. Units will
be mg/L.
}
\description{
Given mass and flow at the monthly basis, the flow-weighted average annual concentration is 
returned. Mass and flow should be monthly data and should be for only a single trace, 
or for one year of one trace. Expect flow to be in acre-ft/month and mass to be in tons.
Return value will be in mg/L
}
\examples{
flow <- rdfSlotToMatrix(rdf,'Powell.Outflow')
mass <- rdfSlotToMatrix(rdf,'Powell.Outflow Salt Mass')
fwaacT1 <- flowWeightedAvgAnnConc(mass[,1], flow[,1]) # repeat for other traces.
}
\seealso{
\code{\link{rdfSlotToMatrix}}
}

