% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_helperFunctions.R
\name{getArrayPctl}
\alias{getArrayPctl}
\title{Get values at the input exceedance levels for the entire array by date}
\usage{
getArrayPctl(rdfXTS, pctlLevels)
}
\arguments{
\item{rdfXTS}{Object returned by \code{\link{rdfSlotToXTS}}}

\item{pctlLevels}{Decimal value(s) for the desired exceedance levels with
0.0<value<1.0}
}
\value{
an XTS object with the selected slot data at the input exceedance
levels
}
\description{
Get values at the input exceedance levels for the entire array by date
}
\examples{
pe <- RWDataPlyr:::rdfSlotToXTS(keyRdf, 'Mead.Pool Elevation')
pe105090 <- RWDataPlyr:::getArrayPctl(pe, c(0.1, 0.5, 0.9))

}
\keyword{internal}
