% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFromRdf.R
\name{getSlots}
\alias{getSlots}
\title{Get and aggregate data from a single rdf file.}
\usage{
getSlots(slotsAndRdf, scenPath)
}
\description{
\code{getSlots} gets all of the slots contained in a single rdf file and aggregates them
as specified by the summary functions in \code{slotsAndRdf}.
}

