% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_helperFunctions.R
\name{getSlotsInRdf}
\alias{getSlotsInRdf}
\title{Returns all slots contained in an rdf file.}
\usage{
getSlotsInRdf(rdf)
}
\arguments{
\item{rdf}{list returned by \code{\link{read.rdf}}}
}
\value{
A vector of strings.  Each string is a slot contained in the rdf list.
}
\description{
\code{getSlotsInRdf} returns a list of all of the slots contained within the rdf list.
}
\examples{
getSlotsInRdf(keyRdf)

}

