% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_helperFunctions.R
\name{getTimeSpan}
\alias{getTimeSpan}
\title{Returns the run period from an rdf.}
\usage{
getTimeSpan(rdf)
}
\arguments{
\item{rdf}{list returned by \code{\link{read.rdf}}}
}
\value{
A named string vector with two elements. The first element, named 'start',
includes the start date of the simulation. The second element, named 'end',
includes the end date of the simulation.
}
\description{
\code{getTimeSpan} Takes a list created by \code{\link{read.rdf}} and returns
the run period from the
}
\examples{
getTimeSpan(keyRdf)

}

