% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_helperFunctions.R
\name{getTraceMin}
\alias{getTraceMin}
\title{Get the annual minimum for each trace}
\usage{
getTraceMin(rdfXTS, yearType)
}
\arguments{
\item{rdfXTS}{XTS object returned by \code{\link{rdfSlotToXTS}}}

\item{yearType}{String 'CY' or 'WY' denoting a Calendar Year (Jan-1 to
Dec-31) or Water Year (Oct-1 to Sep-30)}
}
\value{
an XTS object with the selected slot annual minimum
}
\description{
Get the annual minimum for each trace
}
\examples{
pe <- RWDataPlyr:::rdfSlotToXTS(keyRdf, 'Mead.Pool Elevation')
peWY <- RWDataPlyr:::getTraceMin(pe, 'WY')

}
\keyword{internal}
