% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_helperFunctions.R
\name{getTraceMonthVal}
\alias{getTraceMonthVal}
\title{Get values that meet a month requirement}
\usage{
getTraceMonthVal(rdfXTS, month)
}
\arguments{
\item{rdfXTS}{XTS object returned by \code{\link{rdfSlotToXTS}}}

\item{month}{Integer month(s) as a single value or a vector with 1<=month<=12}
}
\value{
an XTS object with the selected slot data
}
\description{
Get values that meet a month requirement
}
\examples{
\dontrun{
pe <- rdfSlotToXTS(keyRdf, 'Powell.Pool Elevation')
peJanFeb <- getTraceMonthVal(pe, c(1, 2))
}

}
