% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_helperFunctions.R
\name{getTraceSum}
\alias{getTraceSum}
\title{Get the annual sum for each trace}
\usage{
getTraceSum(rdfXTS, yearType)
}
\arguments{
\item{rdfXTS}{XTS object returned by \code{\link{rdfSlotToXTS}}}

\item{yearType}{String 'CY' or 'WY' denoting a Calendar Year (Jan-1 to
Dec-31) or Water Year (Oct-1 to Sep-30)}
}
\value{
an XTS object with the selected slot annual sum
}
\description{
Get the annual sum for each trace
}
\examples{
pe <- RWDataPlyr:::rdfSlotToXTS(keyRdf, 'Mead.Pool Elevation')
peWY <- RWDataPlyr:::getTraceSum(pe, 'WY')

}
\keyword{internal}
