% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWYFromYearMon.R
\name{getWYFromYearmon}
\alias{getWYFromYearmon}
\title{Get the water year from a month-year (yearmon) value}
\usage{
getWYFromYearmon(ym)
}
\arguments{
\item{ym}{An object of class yearmon, or something that can be converted to yearmon}
}
\value{
The water year as a numeric
}
\description{
\code{getWYFromYearmon} returns the water year (assumed to be October - September)
from a yearmon object. October - December of a year, are part of the next water
year and that will be returned.
}
\details{
If the argument is not already a yearmon object, it will attempt to convert it.
This may results in unexpected results. For example, the string \code{"12-1-1906"} can
be converted to a yearmon, however, it will not convert to \code{"Dec 1906"} as
you might desire. It will convert to \code{"Jan 0012"} since it is not a format expected
by \code{zoo::as.yearmon}.
}
\examples{
library(zoo)
getWYFromYearmon(as.yearmon(c("Dec 1906", "Oct 1945", "Jul 1955")))
getWYFromYearmon("2000-11")

}
