% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{is_rdf}
\alias{is_rdf}
\alias{is.rdf}
\title{Test if the object is an rdf}
\usage{
is_rdf(x)

is.rdf(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{rdf} class.
}
\description{
Test if the object is an rdf
}
