% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwd_agg.R
\name{is_rwd_agg}
\alias{is_rwd_agg}
\alias{is.rwd_agg}
\title{Test if the object is a rwd_agg}
\usage{
is_rwd_agg(x)

is.rwd_agg(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{rwd_agg} class.
}
\description{
Test if the object is a rwd_agg
}
