% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot_agg_list.R
\name{is_slot_agg_list}
\alias{is_slot_agg_list}
\alias{is.slot_agg_list}
\title{Test if the object is a slot_agg_list}
\usage{
is_slot_agg_list(x)

is.slot_agg_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{slot_agg_list} class.
}
\description{
Test if the object is a slot_agg_list
}
