% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keyRdf}
\alias{keyRdf}
\title{Example rdf file with monthly data.}
\format{
A multi level list. \code{keyRdf$meta} provides a description
of the RiverWare run used to generate this data.
}
\source{
Bureau of Reclamation, 2016
}
\usage{
keyRdf
}
\description{
An example of an rdf file that has already been read into R via
\code{\link[=read.rdf]{read.rdf()}}. This example contains 39 slots, at the monthly
timestep for 11 years and 25 runs. Slots include pool elevation,
flow, and flags. Use this with \code{\link[=rdf_slot_names]{rdf_slot_names()}} or
\code{\link[=rdf_get_slot]{rdf_get_slot()}} to use the data.
}
\keyword{datasets}
