% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listSlots}
\alias{listSlots}
\title{List all slots contained in an rdf file (list returned by read.rdf)}
\usage{
listSlots(rdf)
}
\arguments{
\item{rdf}{list returned by read.rdf}
}
\value{
A vector of strings.  Each string is a slot contained in the rdf list.
}
\description{
List all slots contained in an rdf file (list returned by read.rdf)
}
\examples{
zz <- read.rdf('KeySlots.rdf')
listSlots(zz)
}

