% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_helperFunctions.R
\name{rdf_get_slot}
\alias{rdf_get_slot}
\alias{rdfSlotToMatrix}
\title{Get a slot out of an rdf object}
\usage{
rdf_get_slot(rdf, slot)

rdfSlotToMatrix(rdf, slot)
}
\arguments{
\item{rdf}{An \code{rdf} object.}

\item{slot}{Character slot name that exists in the \code{rdf}.}
}
\value{
A matrix with traces as columns and timesteps as rows.
}
\description{
\code{rdf_get_slot()} gets a slot from an \code{rdf} object and creates a matrix with
rows indexing through time and columns indexing over traces.
}
\section{Functions}{
\itemize{
\item \code{rdfSlotToMatrix}: Deprecated version of \code{rdf_get_slot()}
}}

\examples{
pe <- rdf_get_slot(keyRdf, "Mead.Pool Elevation")

}
