% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_helperFunctions.R
\name{rdf_get_timespan}
\alias{rdf_get_timespan}
\title{Returns the simulation timespan from an rdf}
\usage{
rdf_get_timespan(rdf)
}
\arguments{
\item{rdf}{An \code{rdf} object (likely from \code{\link[=read.rdf]{read.rdf()}}).}
}
\value{
A named character vector with two elements. The first element, named
\code{"start"}, includes the start date of the simulation. The second element,
named \code{"end"}, includes the end date of the simulation.
}
\description{
\code{rdf_get_timespan()} gets the simulation timespan from an \code{rdf} object.
}
\examples{
rdf_get_timespan(keyRdf)

}
