% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_helperFunctions.R
\name{rdf_slot_names}
\alias{rdf_slot_names}
\alias{getSlotsInRdf}
\title{Returns all slots contained in an rdf file.}
\usage{
rdf_slot_names(rdf)

getSlotsInRdf(rdf)
}
\arguments{
\item{rdf}{An \code{rdf} object.}
}
\value{
A character vector.
}
\description{
\code{rdf_slot_names()} returns a character vector of all slots contained within
an rdf object.
}
\section{Functions}{
\itemize{
\item \code{getSlotsInRdf}: Deprecated version of \code{rdf_slot_names()}
}}

\examples{
rdf_slot_names(keyRdf)

}
\seealso{
\code{\link[=read.rdf]{read.rdf()}}
}
