% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_helperFunctions.R
\name{rdfSlotToMatrix}
\alias{rdfSlotToMatrix}
\title{Get one slot out of an rdf list and put it in matrix form.}
\usage{
rdfSlotToMatrix(rdf, slot)
}
\arguments{
\item{rdf}{list returned by \code{\link{read.rdf}}}

\item{slot}{string of slot name that exists in \code{rdf} that will be converted to a matrix}
}
\description{
\code{rdfSlotToMatrix} Takes a list created by \code{\link{read.rdf}} and converts the nested slot values over
multiple traces into a matrix with rows indexing through time and columns
indexing over traces.
}
\examples{
pe <- rdfSlotToMatrix(keyRdf, 'Mead.Pool Elevation')

}
