% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_helperFunctions.R
\name{rdfSlotToXTS}
\alias{rdfSlotToXTS}
\title{Get one slot out of an rdf list and put it in an XTS object}
\usage{
rdfSlotToXTS(rdf, slot)
}
\arguments{
\item{rdf}{list returned by \code{\link{read.rdf}}}

\item{slot}{string of slot name that exists in \code{rdf} that will be
converted to a matrix}
}
\value{
an XTS object with the selected slot data
}
\description{
\code{rdfSlotToXTS} Takes a list created by \code{\link{read.rdf}} and
convert the nested slot values over the multiple traces into an XTS array
indexing over traces.
}
\examples{
pe <- RWDataPlyr:::rdfSlotToXTS(keyRdf, 'Mead.Pool Elevation')

}
\keyword{internal}
