% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfToTbl.R
\name{rdfToTbl}
\alias{rdfToTbl}
\title{Convert RDF to a Tibble}
\usage{
rdfToTbl(rdf, scenario = NULL)
}
\arguments{
\item{rdf}{An rdf list returned from \code{\link{read.rdf2}}.}

\item{scenario}{An optional parameter, that if it is not \code{NULL} (default)
will be added to the tibble as another variable. Typically a string, but it is
not coerced to a string.}
}
\value{
A tibble with additional attributes from the rdf list
}
\description{
\code{rdfToTbl} converts an rdf list to a tibble (data.frame).
}
\details{
The rdf list is converted to a "long" data frame, and then converted to a 
\code{\link[tibble]{tibble}}. All of the \code{meta} entries into the rdf list
are stored as attribures in the returned tibble.
}
\examples{
t1 <- rdfToTbl(keyRdf)
t2 <- rdfToTbl(sysRdf, scenario = "ISM1988_2014,2007Dems,IG,2002")

}
\seealso{
\code{\link{read.rdf2}}, \code{\link{read.rdf}}
}
