% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwd_agg_helpers.R
\name{read_rwd_agg}
\alias{read_rwd_agg}
\title{Read in a rwd_agg file}
\usage{
read_rwd_agg(file)
}
\arguments{
\item{file}{The csv file to be read in and converted}
}
\description{
\code{read_rwd_agg()} reads in a csv file and creates a \link{rwd_agg} object.
Therefore, if the csv file is not properly formatted to contain the correct
information for a \link{rwd_agg} object, it will fail. \code{\link[=rwd_agg_template]{rwd_agg_template()}} will
create a blank template file for the user to fill in, which has the correct
headers.
}
\examples{
read_rwd_agg( 
  system.file(
    "extdata/rwd_agg_files/passing_aggs.csv", 
    package = "RWDataPlyr"
  )
)

}
\seealso{
\code{\link[=rwd_agg_template]{rwd_agg_template()}}
}
