% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRwCsv.R
\name{readRwCsv}
\alias{readRwCsv}
\title{Read in RiverWare Data}
\usage{
readRwCsv(file)
}
\arguments{
\item{file}{The name of the file which the data are to be read from. Either an
absolute or relative path.}
}
\value{
A data frame (\code{\link[base]{data.frame}}) containing a 
representation of the data in the file.
}
\description{
Read in a CSV file created from RiverWare. If the CSV file does not contain 
column names that RiverWare always uses (see Details), then it assumes that the CSV file
was not created from RiverWare and throws an error. It also removes spaces
from the column names, and adjusts the \code{Object.Slot} and \code{Slot Value}
columns to be \code{ObjectSlot} and \code{Value}, respectively.
}
\details{
The required column names are: \code{Run Number}, \code{Trace Number}, 
\code{Object.Slot}, \code{Timestep}, \code{Slot Value}. See the CSV output 
section of the 
\href{http://www.riverware.org/PDF/RiverWare/documentation/MRM.pdf}{RiverWare documentation}
for more information on the other optional column names.

This function uses \code{data.table::\link[data.table]{fread}()} to read in 
the CSV file, and forces it to expect a CSV file, expect headers, and return
a \code{data.frame}.
}
\examples{
zz <- readRwCsv(system.file(
  "extdata/Scenario/ISM1988_2014,2007Dems,IG,Most",
  "KeySlots.csv",
  package = "RWDataPlyr"
))

}
\seealso{
\code{\link{read.rdf}}
}
