% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWDataPlyr.R
\docType{package}
\name{RWDataPlyr-package}
\alias{RWDataPlyr}
\alias{RWDataPlyr-package}
\title{RWDataPlyr: Read and manipulate data from RiverWareTM}
\description{
A tool to read and manipulate data generated from RiverWare(TM)
(riverware.org) simulations. RiverWare generates data in "rdf" format. This
package provides an interface to read, combine data from multiple
scenarios, and summarize the data.
}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

Authors:
\itemize{
  \item Cameron Bracken
}

}
\keyword{internal}
