% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scen_data}
\alias{scen_data}
\title{Example aggregated scenario data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 720 rows and 6 columns.
}
\usage{
scen_data
}
\description{
An example of the tbl_df returned by \code{\link[=rw_scen_aggregate]{rw_scen_aggregate()}} containing two
scenarios of data.
}
\keyword{datasets}
