% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sumMonth2Annual}
\alias{sumMonth2Annual}
\title{Sums monthly trace data into annual values.}
\usage{
sumMonth2Annual(matrixToSum, multFactor = 1)
}
\arguments{
\item{matrixToSum}{The monthly trace data (months by traces) that will be summed.}

\item{multFactor}{A factor the annual sum will be multiplied by.  Can be used to convert from flow to volume.}
}
\value{
The annual sums as a matrix (years by traces).
}
\description{
Sums monthly trace data into annual values.
}
\examples{
zz <- read.rdf('KeySlots.rdf')
zz <- rdfSlotToMatrix(zz, 'Powell.Outflow')
annualRelease <- sumMonth2Annual(zz) # returns in original units, e.g., acre-ft
annualRelease <- sumMonth2Annual(zz,.001) # returns in scaled units, e.g., kaf
}

