% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSlotAggList.R
\name{createSlotAggList}
\alias{createSlotAggList}
\title{Creates a list for use by \code{\link{getDataForAllScens}}.}
\usage{
createSlotAggList(iData)
}
\arguments{
\item{iData}{Either an N x 4 character matrix or a character with an absolute
or relative path to a csv file.}
}
\description{
Deprecated: please use \code{\link[=slot_agg_list]{slot_agg_list()}} instead, which returns the same
list, but now as a "slot_agg_list" object.
}
