% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFromRdf.R
\name{getAndProcessAllSlots}
\alias{getAndProcessAllSlots}
\title{Get and aggregate data from rdf file(s) for one scenario.}
\usage{
getAndProcessAllSlots(scenPath, slotAggList)
}
\arguments{
\item{scenPath}{A relative or absolute path to the scenario folder.}

\item{slotAggList}{The slot aggregation list. A list containing the slots that will be 
imported and aggregated, the aggregation method(s) to use, and the rdf files that 
contain the slots. See \code{\link{createSlotAggList}}.}
}
\description{
\code{getAndProcessAllSlots} gets data for a single scenario.  The slots from each
rdf are processed and aggregated together.
}
\seealso{
\code{\link{getDataForAllScens}}
}

