% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFromRdf.R
\name{getAndProcessAllSlots}
\alias{getAndProcessAllSlots}
\title{Get and aggregate data from rdf file(s) for one scenario.}
\usage{
getAndProcessAllSlots(scenPath, slotAggList, findAllSlots)
}
\arguments{
\item{scenPath}{A relative or absolute path to the scenario folder.}

\item{slotAggList}{The slot aggregation list. A list containing the slots that will be 
imported and aggregated, the aggregation method(s) to use, and the rdf files that 
contain the slots. Either created by calling \code{\link{createSlotAggList}} with a specified
set of slots, or a list of lists with each entry containing an rdf file and the keyword 
'all' for the slots, e.g., \code{list(list(rdf = 'KeySlots.rdf',slots = 'all'))}, which will
return all of the slots found in an rdf file. If this option is used, the code will return
monthly, or annual data, i.e., no aggregation methods will be applied to the data in the rdf
file.}

\item{findAllSlots}{Boolean; if \code{TRUE} (default), then the function will
abort if it cannot find a particular slot. If \code{FALSE}, then the function
will continue, even if a slot cannot be found. If a slot is not found, then the
function will return \code{-99} for the Trace, Year, and Value.}
}
\description{
\code{getAndProcessAllSlots} gets data for a single scenario.  The slots from each
rdf are processed and aggregated together.
}
\seealso{
\code{\link{getDataForAllScens}}
}
\keyword{internal}
