% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDataFromRdf.R
\name{getAndProcessAllSlots}
\alias{getAndProcessAllSlots}
\title{Get and aggregate data from rdf file(s) for one scenario.}
\usage{
getAndProcessAllSlots(scenPath, slotsAndRdf)
}
\arguments{
\item{scenPath}{A relative or absolute path to the scenario folder.}
}
\description{
\code{getAndProcessAllSlots} gets data for a single scenario.  The slots from each
rdf are processed and aggregated together.
}
\seealso{
\code{\link{getDataForAllScens}}
}

