% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFromRdf.R
\name{getSlots}
\alias{getSlots}
\title{Get and aggregate data from a single rdf file.}
\usage{
getSlots(slotAggList, scenPath)
}
\arguments{
\item{slotAggList}{The slot aggregation list. A list containing the slots that will be 
imported and aggregated, the aggregation method(s) to use, and the rdf files that 
contain the slots. See \code{\link{createSlotAggList}}.}

\item{scenPath}{A relative or absolute path to the scenario folder.}
}
\description{
\code{getSlots} gets all of the slots contained in a single rdf file and aggregates them
as specified by the summary functions in \code{slotAggList}.
}
\keyword{internal}
