% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_helperFunctions.R
\name{getTraceMax}
\alias{getTraceMax}
\title{Get the annual maximum for each trace}
\usage{
getTraceMax(rdfXTS, yearType)
}
\arguments{
\item{rdfXTS}{XTS object returned by \code{\link{rdfSlotToXTS}}}

\item{yearType}{String 'CY' or 'WY' denoting a Calendar Year (Jan-1 to
Dec-31) or Water Year (Oct-1 to Sep-30)}
}
\value{
an XTS object with the selected slot annual maximum
}
\description{
Get the annual maximum for each trace
}
\examples{
\dontrun{
pe <- rdfSlotToXTS(keyRdf, 'Powell.Pool Elevation')
peWY <- getTraceMax(pe, 'WY')
}
}
