% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwd_agg.R, R/slot_agg_list.R
\name{is.rwd_agg}
\alias{is.rwd_agg}
\alias{is_slot_agg_list}
\alias{is.slot_agg_list}
\title{Test if the object is a slot_agg_list}
\usage{
is.rwd_agg(x)

is_slot_agg_list(x)

is.slot_agg_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{slot_agg_list} class.
}
\description{
Test if the object is a slot_agg_list
}
