% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listSlots}
\alias{listSlots}
\title{List all slots contained in an rdf file.}
\usage{
listSlots(rdf)
}
\arguments{
\item{rdf}{list returned by \code{\link{read.rdf}}}
}
\value{
A vector of strings.  Each string is a slot contained in the rdf list.
}
\description{
\code{listSlots} returns a list of all of the slots contained within the rdf list.
}
\examples{
zz <- read.rdf('KeySlots.rdf')
listSlots(zz)
}

