% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rdfSlotToMatrix}
\alias{rdfSlotToMatrix}
\title{Takes a list created by \code{\link{read.rdf}} and converts the nested slot values over
multiple traces into a matrix with rows indexing through time and columns
indexing over traces}
\usage{
rdfSlotToMatrix(rdf, slot)
}
\arguments{
\item{rdf}{list returned by \code{\link{read.rdf}}}

\item{slot}{string of slot name that exists in \code{rdf} that will be converted to a matrix}
}
\description{
Takes a list created by \code{\link{read.rdf}} and converts the nested slot values over
multiple traces into a matrix with rows indexing through time and columns
indexing over traces
}
\examples{
zz <- read.rdf('KeySlots.rdf')
pe <- rdfSlotToMatrix(zz, 'Powell.Pool Elevation')
}

