% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rdf.R
\name{read.rdf}
\alias{read.rdf}
\title{Read an rdf file into R.}
\usage{
read.rdf(file)
}
\arguments{
\item{file}{The input rdf file that will be read into R.}
}
\value{
A multi-level list containing all metadata and slot data in the original rdf file.
}
\description{
\code{read.rdf} reads an rdf file into R and formats it as a multi-level list containing
all of the metadata included in the rdf file.  Rdf files are generated by RiverWare
and are documented in the \href{http://riverware.org/PDF/RiverWare/documentation/}{RiverWare documentation}.
}
\examples{
zz <- read.rdf(system.file('extdata/Scenario/DNF,CT,IG', "KeySlots.rdf", package = "RWDataPlot"))

}

