% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_to_tbl.R
\name{rw_rdf_to_tbl}
\alias{rw_rdf_to_tbl}
\title{Convert RDF to a Tibble}
\usage{
rw_rdf_to_tbl(rdf, scenario = NULL)
}
\arguments{
\item{rdf}{An rdf list returned from \code{\link[=read.rdf]{read.rdf()}}.}

\item{scenario}{An optional parameter, that if it is not \code{NULL} (default)
will be added to the tibble as another variable. Typically a string, but it
is not coerced to a string.}
}
\value{
A tibble with additional attributes from the rdf list
}
\description{
\code{rw_rdf_to_tbl()} converts an rdf list to a tibble (data.frame).
}
\details{
The rdf list is converted to a "long" data frame, and then converted to a
\code{\link[tibble:tibble]{tibble::tibble()}}. All of the \code{meta} entries into the rdf list
are stored as attribures in the returned tibble.
}
\examples{
t1 <- rw_rdf_to_tbl(keyRdf)
t2 <- rw_rdf_to_tbl(sysRdf, scenario = "ISM1988_2014,2007Dems,IG,2002")

}
\seealso{
\code{\link[=read.rdf]{read.rdf()}}
}
