% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_to_tbl.R
\name{rw_rdf_to_tbl}
\alias{rw_rdf_to_tbl}
\title{Convert rdf to a tibble}
\usage{
rw_rdf_to_tbl(rdf, scenario = NULL, keep_cols = FALSE, add_ym = TRUE)
}
\arguments{
\item{rdf}{An rdf list returned from \code{\link[=read.rdf]{read.rdf()}}.}

\item{scenario}{An optional parameter, that if it is not \code{NULL} (default)
will be added to the tibble as another variable. Typically a string, but it
is not coerced to a string.}

\item{keep_cols}{Either boolean, or a character vector of column names to
keep in the returned tibble. The values of \code{keep_cols} work as follows:
\itemize{
\item \code{FALSE} (default) only includes the defaults columns: \code{Timestep},
\code{TraceNumber}, \code{ObjectSlot}, and \code{Value}. \code{Scenario} is also returned if
\code{scenario} is specified.
\item \code{TRUE}, all columns are returned.
\item A character vector, e.g., \code{c("ObjectName", "Units")}, allows the user to
include other columns that are not always required, in addition to the
"default" set of columns. If any of the values in \code{keep_cols} are not
found, a warning will post, but all other columns will be returned.
}}

\item{add_ym}{Boolean that controls whether or not \code{Year} and \code{Month} columns
are added to the returned tibble. If \code{TRUE} (default), they will be added,
and if \code{FALSE} they will not be added. They are constructed from the dates
in the \code{Timestep} column.}
}
\value{
A tibble with additional attributes from the rdf list.
}
\description{
\code{rw_rdf_to_tbl()} converts an rdf list to a tibble (data.frame).
}
\details{
The rdf list is converted to a data frame, and then converted to a
\code{\link[tibble:tibble]{tibble::tibble()}}. All of the \code{meta} entries into the rdf list
are stored as attributes in the returned tibble. These attributes are:
\code{mrm_config_name}, \code{owner}, \code{description}, \code{create_date}, and \code{n_traces}.
}
\examples{
rdftbl <- rw_rdf_to_tbl(keyRdf)
# same as previous, except you do not want "Year" and "Month" columns
rdftbl <- rw_rdf_to_tbl(keyRdf, add_ym = FALSE)
# but you do want to keep the object name seperately:
rdftbl <- rw_rdf_to_tbl(keyRdf, add_ym = FALSE, keep_cols = "Object")
rdftbl <- rw_rdf_to_tbl(sysRdf, scenario = "ISM1988_2014,2007Dems,IG,2002")

}
\seealso{
\code{\link[=read.rdf]{read.rdf()}}
}
