% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWDataPlyr.R
\docType{package}
\name{RWDataPlyr-package}
\alias{RWDataPlyr}
\alias{RWDataPlyr-package}
\title{RWDataPlyr: Read and Manipulate Data from 'RiverWare'}
\description{
A tool to read and manipulate data generated from 'RiverWare'(TM) 
    <http://www.riverware.org/> simulations. 'RiverWare' and 'RiverSMART' 
    generate data in "rdf", "csv", and "nc" format. This package provides an 
    interface to read, aggregate, and summarize data from one or more 
    simulations in a 'dplyr' pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BoulderCodeHub/RWDataPlyr}
  \item Report bugs at \url{https://github.com/BoulderCodeHub/RWDataPlyr/issues}
}

}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

Authors:
\itemize{
  \item Cameron Bracken
}

}
\keyword{internal}
