% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWDataPlyr.R
\docType{package}
\name{RWDataPlyr-package}
\alias{RWDataPlyr}
\alias{RWDataPlyr-package}
\title{RWDataPlyr: Read and manipulate data from RiverWare}
\description{
A tool to read and manipulate data generated from RiverWare(TM)
(riverware.org) simulations. RiverWare and RiverSMART generate data in
"rdf", "csv", and "nc" format. This package provides an interface to read,
aggregate, and summarize data from one or more simulations.
}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

Authors:
\itemize{
  \item Cameron Bracken
}

}
\keyword{internal}
