% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwslot_aggregations.R
\name{rwslot_annual_sum}
\alias{rwslot_annual_sum}
\alias{sumMonth2Annual}
\alias{getMinAnnValue}
\alias{getMaxAnnValue}
\alias{flowWeightedAvgAnnConc}
\title{Simple aggregation functions for monthly matrix data}
\usage{
rwslot_annual_sum(rwslot, multFactor = 1)

sumMonth2Annual(rwslot, multFactor = 1)

getMinAnnValue(rwslot)

getMaxAnnValue(rwslot)

flowWeightedAvgAnnConc(mass, flow)
}
\arguments{
\item{rwslot}{A matrix (months by traces) such as that returned by
\code{\link[=rdf_get_slot]{rdf_get_slot()}}. Function will error if the number of rows in \code{rwslot} is
not divisible by 12, i.e., the data must be monthly for a full consecutive
year.}

\item{multFactor}{A factor the annual sum will be multiplied by.  Can be used
to convert from flow to volume, or to scale all results in another manor.}

\item{mass}{A vector of one trace worth of data, or one year of one trace.
Units should be in tons.}

\item{flow}{A vector of one trace worth of data, or one year of one trace.
Units should be in acre-ft/month.}
}
\value{
\code{rwslot_annual_sum()} returns the annual sum as a matrix (years by
traces).

\code{getMinAnnValue()} returns a matrix (years by traces) with the
maximum annual value for each year and trace.

\code{getMaxAnnValue()} returns a matrix (years by traces) with the
maximum annual value for each year and trace.

\code{flowWeightedAvgAnnConc()} returns a vector of yearly data of the
flow-weighted average annual concentration. Units will be mg/L.
}
\description{
\code{rwslot_annual_sum()} takes a matrix containing monthly data (months by
traces), and sums the monthly data into annual data.  Returns a years by
traces matrix.

\code{getMinAnnValue()} finds the minimum annual value for all years and traces.

\code{getMaxAnnValue()} finds the maximum annual value for all years and traces.

\code{flowWeightedAvgAnnConc()} calculates the flow-weighted average annual
concentration. Given mass and flow at the monthly basis, the flow-weighted
average annual concentration is returned. Mass and flow should be monthly
data and should be for only a single trace, or for one year of one trace.
Expect flow to be in acre-ft/month and mass to be in tons. Return value will
be in mg/L.
}
\section{Functions}{
\itemize{
\item \code{sumMonth2Annual}: Deprecated version of \code{rwslot_annual_sum()}.
}}

\examples{
zz <- rdf_get_slot(keyRdf, 'Powell.Outflow')

# returns in original units, e.g., acre-ft
annualTotVal <- rwslot_annual_sum(zz)

# returns in scaled units, e.g., kaf
annualTotVal <- rwslot_annual_sum(zz, 0.001) 

pe <- rdf_get_slot(keyRdf,'Mead.Pool Elevation')
peMax <- getMinAnnValue(pe)

pe <- rdf_get_slot(keyRdf,'Mead.Pool Elevation')
peMax <- getMaxAnnValue(pe)

\dontrun{
flow <- rdf_get_slot(rdf,'Powell.Outflow')
mass <- rdf_get_slot(rdf,'Powell.Outflow Salt Mass')
# repeat for other traces as necessary:
fwaacT1 <- flowWeightedAvgAnnConc(mass[,1], flow[,1]) 
}

}
\seealso{
\code{\link[=rdf_get_slot]{rdf_get_slot()}}
}
