% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_agg_scen_data.R
\name{rwtbl_summarise_all}
\alias{rwtbl_summarise_all}
\title{\code{rwtbl_summarise_all()} processes a \code{slot_agg_matrix} for a single scenario.}
\usage{
rwtbl_summarise_all(slot_agg_matrix, scen_dir = ".", scenario = NULL,
  keep_cols = FALSE)
}
\description{
\code{rwtbl_summarise_all()} processes a \code{slot_agg_matrix} for a single scenario.
}
