% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{sumMonth2Annual}
\alias{sumMonth2Annual}
\title{Sums monthly trace data into annual values.}
\usage{
sumMonth2Annual(matrixToSum, multFactor = 1)
}
\arguments{
\item{matrixToSum}{The monthly trace data (months by traces) that will be summed.}

\item{multFactor}{A factor the annual sum will be multiplied by.  Can be used to convert from flow to volume.}
}
\value{
The annual sums as a matrix (years by traces).
}
\description{
\code{sumMonth2Annual} takes a matrix containing monthly data (months by traces), 
sum the monthly data into annual data.  Returns a years by traces matrix.
}
\examples{
zz <- rdfSlotToMatrix(keyRdf, 'TotVal.Powell')
annualTotVal <- sumMonth2Annual(zz) # returns in original units, e.g., acre-ft
annualTotVal <- sumMonth2Annual(zz,.001) # returns in scaled units, e.g., kaf
}

